+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add vkBasalt for use with GOverlay.
 - Change the 'linux-neptune' package to be 'linux-steamos'.
 - Change Pacman to use 'wget' for more reliable downloads on slower internet.
+- Add the ability to build winesapOS using a proxy.
 
 ## 3.1.1 - The Boot Fix Update